:recycle: Notionからデータを取ってくる場所を変更
